requirements.txt
chardet==4.0.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
idna==2.10
importlib-resources==6.1.1
Jinja2==3.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mysql-connector-python==8.2.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==8.4.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
zipp==3.17.0
